[🚀 사이클2 - 미션 (예약 변경/취소와 에러 처리)] 주니(이해선) 미션 제출합니다.  (#433)
* docs: README.md 작성

* feat: `@RestControllerAdvice`로 전역 에외 처리 추가

* feat: 500 에러가 사용자에게 노출되지 않도록 처리 및 애러 응답 JSON 형식으로 통일

* refactor: data.sql 예약 날짜 수정

* refactor: 오늘 날짜에서 이미 지난 시간 예약할 수 없도록 변경

* feat: 브라우저에서 에러 발생 시 사용자에게 메세지 표시

* docs: 예약 API 명세서에 예약 날짜 및 시간 변경 API 추가

* feat: 본인의 예약 날짜·시간 변경 API 추가

* feat: 예약 변경 화면 구현 및 에러 케이스 처리

* test: ReservationService 예약 변경 관련 테스트 추가

* test: UserReservationTest 예약 날짜 및 시간 변경 테스트 추가

* refactor: 에러 응답 코드를 도메인 의미 기반으로 변경

* refactor: 예약 페이지 아이콘 정렬 수정

* refactor: Request/Response 클래스 domain 패키지에서 분리

* refactor: 생성 시간 검증 로직 추가

* refactor: JsonFormat 어노테이션을 JacksonConfig 전역 설정으로 분리

* feat: 지나간 시간에 대한 검증 로직 추가

* test: 지나간 시간에 대한 검증 로직 테스트

* refactor: 에러 로그 명확하게 변경

* feat: 취소 케이스에 대한 검증 로직 추가

* feat: 이미 지난 시간에 대한 예약 업데이트 검증 로직 추가

* refactor: 메서드명 통일성 있게 변경

* refactor: 검증 메세지 형식 변경

* refactor: ReferencedDataException 메세지 사용자 친화적으로 변경

* test: 예약 예외 상황에 대한 테스트 추가

* test: 에외 관련 테스트 추가 및 보완

+++ src/main/java/roomescape/service/ReservationService.java
@@ -2,10 +2,11 @@ package roomescape.service;
 
 import org.springframework.stereotype.Service;
 import roomescape.domain.reservation.Reservation;
-import roomescape.domain.reservation.ReservationRequest;
-import roomescape.domain.reservation.ReservationResponse;
+import roomescape.dto.reservation.ReservationRequest;
+import roomescape.dto.reservation.ReservationResponse;
 import roomescape.domain.reservationtime.ReservationTime;
 import roomescape.domain.theme.Theme;
+import roomescape.exception.InvalidReservationException;
 import roomescape.exception.ReservationAlreadyExistException;
 import roomescape.exception.ReservationNotFoundException;
 import roomescape.exception.ReservationTimeNotFoundException;
@@ -16,6 +17,7 @@ import roomescape.repository.ReservationUpdatingDao;
 import roomescape.repository.ThemeQueryingDao;
 
 import java.time.LocalDate;
+import java.time.LocalTime;
 import java.util.List;
 import java.util.Optional;
 
@@ -61,7 +63,11 @@ public class ReservationService {
                 .orElseThrow(() -> new ThemeNotFoundException(reservationReq.themeId()));
 
         if (reservationReq.date().isBefore(LocalDate.now())) {
-            throw new IllegalArgumentException("현재보다 이전의 날짜는 예약할 수 없습니다.");
+            throw new InvalidReservationException();
+        }
+
+        if (reservationReq.date().isEqual(LocalDate.now()) && reservationTimeById.getStartAt().isBefore(LocalTime.now())) {
+            throw new InvalidReservationException();
         }
 
         Optional<Reservation> savedReservation = reservationQueryingDao.findReservationByThemeAndDateAndTime(themeById.getId(), reservationReq.date(), reservationTimeById.getId());
@@ -71,10 +77,43 @@ public class ReservationService {
 
         Reservation reservation = new Reservation(reservationReq.name(), reservationReq.date(), reservationTimeById, themeById);
         Long generatedId = reservationUpdatingDao.insert(reservation);
-        return ReservationResponse.from(reservation.reservationWithId(generatedId));
+        return ReservationResponse.from(reservation.withReservationId(generatedId));
+    }
+
+    public ReservationResponse update(Long id, ReservationRequest reservationReq) {
+        Reservation existedReservation = reservationQueryingDao.findReservationById(id)
+                .orElseThrow(() -> new ReservationNotFoundException(id));
+        ReservationTime newTime = reservationTimeQueryingDao.findReservationTimeById(reservationReq.timeId())
+                .orElseThrow(() -> new ReservationTimeNotFoundException(reservationReq.timeId()));
+
+        if (reservationReq.date().isBefore(LocalDate.now())) {
+            throw new InvalidReservationException();
+        }
+
+        if (reservationReq.date().isEqual(LocalDate.now()) && newTime.getStartAt().isBefore(LocalTime.now())) {
+            throw new InvalidReservationException();
+        }
+
+        Optional<Reservation> duplicateReservation = reservationQueryingDao.findReservationByThemeAndDateAndTime(existedReservation.getTheme().getId(), reservationReq.date(), newTime.getId());
+        if (duplicateReservation.isPresent()) {
+            throw new ReservationAlreadyExistException();
+        }
+        Reservation updatedReservation = existedReservation.withUpdatedDateAndTime(reservationReq.date(), newTime);
+        reservationUpdatingDao.update(id,  updatedReservation);
+        return ReservationResponse.from(updatedReservation);
     }
 
     public void delete(Long id) {
+        Reservation reservation = reservationQueryingDao.findReservationById(id)
+                .orElseThrow(() -> new ReservationNotFoundException(id));
+
+        if (reservation.getDate().isBefore(LocalDate.now())) {
+            throw new InvalidReservationException();
+        }
+        if (reservation.getDate().isEqual(LocalDate.now()) && reservation.getTime().getStartAt().isBefore(LocalTime.now())) {
+            throw new InvalidReservationException();
+        }
+
         reservationUpdatingDao.delete(id);
     }
 }
